requirements.txt
# Runtime dependencies
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
psycopg[binary]>=3.2.0
pydantic>=2.8.0
kafka-python>=2.0.2
faker>=28.0.0
pandas>=2.2.0
pyarrow>=17.0.0
pyspark>=3.5.0

# Development / testing dependencies (optional)
pytest>=8.3.0
httpx>=0.27.0
ruff>=0.5.0
